# src/main/resources/swagger-ui/swagger.yaml
openapi: 3.0.3

info:
  title: Feedback API
  version: 1.0.0
  description: API for creating, retrieving, updating, and deleting feedback records.

tags:
  - name: Feedback
    description: Operations related to feedback management.

paths:
  /feedback:
    post:
      summary: Create a new feedback
      description: Creates a new feedback record.
      operationId: createFeedback
      tags:
        - Feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      responses:
        '201':
          description: Feedback successfully created.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - Feedback
      summary: List feedbacks
      description: Retrieves a paginated list of feedbacks.
      operationId: listFeedbacks
      parameters:
        - name: description
          in: query
          required: false
          description: Feedback text.
          schema:
            type: string
            example: Excellent service

        - name: urgencyType
          in: query
          required: false
          description: Feedback urgency level.
          schema:
            $ref: '#/components/schemas/UrgencyTypeEnum'

        - name: limit
          in: query
          required: false
          description: Maximum number of records per page.
          schema:
            type: integer
            example: 10

        - name: offset
          in: query
          required: false
          description: Pagination offset.
          schema:
            type: integer
            example: 0

      responses:
        '200':
          description: List of feedbacks found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedFeedbackResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /feedback/{id}:
    get:
      tags:
        - Feedback
      summary: Get feedback by ID
      description: Retrieves a feedback record by its unique identifier.
      operationId: getFeedbackById
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Feedback found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags:
        - Feedback
      summary: Update feedback
      description: Updates an existing feedback record.
      operationId: updateFeedback
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      responses:
        '204':
          description: Feedback successfully updated.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Feedback
      summary: Delete feedback by ID
      description: Deletes a feedback record by its ID.
      operationId: deleteFeedback
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '204':
          description: Feedback successfully deleted.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  schemas:
    FeedbackRequest:
      type: object
      required:
        - description
        - rating
        - urgencyType
      properties:
        description:
          type: string
          description: Feedback text.
          example: Excellent service.
        rating:
          type: number
          description: Rating assigned to the feedback (0 to 10).
          example: 8
        urgencyType:
          $ref: '#/components/schemas/UrgencyTypeEnum'

    FeedbackResponse:
      type: object
      properties:
        id:
          type: string
          example: 652ff3a9b1c2d40012ab45de
        description:
          type: string
          example: Excellent service.
        rating:
          type: number
          example: 8
        urgencyType:
          $ref: '#/components/schemas/UrgencyTypeEnum'
        sentDate:
          type: string
          format: date-time
          example: 2025-10-22T20:30:00Z

    UrgencyTypeEnum:
      type: string
      description: Urgency level associated with the feedback.
      enum:
        - LOW
        - MEDIUM
        - HIGH
        - URGENT
      example: MEDIUM

    PagedFeedbackResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/FeedbackResponse'
        totalElements:
          type: integer
          format: int64
          example: 120
        limit:
          type: integer
          example: 10
        offset:
          type: integer
          example: 0

    ApiError:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Invalid parameters.
        path:
          type: string
          example: /feedback
        timestamp:
          type: string
          format: date-time
          example: 2025-10-11T10:30:00Z

  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Feedback ID (required for update and delete operations).

  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

    Unauthorized:
      description: Authentication failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

    Forbidden:
      description: Access denied.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'